+++ code/renderers/vue3/src/__tests__/Button.vue
@@ -1,58 +1,16 @@
-<template>
-  <button
-    type="button"
-    :class="classes"
-    :style="style"
-    @change="emit('myChangeEvent', 0)"
-    @click="emit('myClickEvent', 0)"
-  >
-    {{ label }}
-  </button>
-</template>
-
-<script lang="ts" setup>
-import './button.css';
-import { computed } from 'vue';
-
-const props = withDefaults(
-  defineProps<{
-    /**
-     * The label of the button
-     */
-    label: string;
-    /**
-     * Whether the button is disabled
-     */
-    disabled: boolean;
-    /**
-     * primary or secondary button
-     */
-    primary?: boolean;
-    /**
-     * size of the button
-     */
-    size?: 'small' | 'medium' | 'large';
-    /**
-     * background color of the button
-     */
-    backgroundColor?: string;
-  }>(),
-  { primary: false }
-);
+<script setup lang="ts">
+defineProps<{ disabled: boolean; label: string }>();
 
 const emit = defineEmits<{
   (e: 'myChangeEvent', id: number): void;
   (e: 'myClickEvent', id: number): void;
 }>();
-
-const classes = computed(() => ({
-  'storybook-button': true,
-  'storybook-button--primary': props.primary,
-  'storybook-button--secondary': !props.primary,
-  [`storybook-button--${props.size || 'medium'}`]: true,
-}));
-
-const style = computed(() => ({
-  backgroundColor: props.backgroundColor,
-}));
 </script>
+
+<template>
+  <button :disabled="disabled" @change="emit('myChangeEvent', 0)" @click="emit('myClickEvent', 0)">
+    {{ label }}
+  </button>
+</template>
+
+<style scoped></style>
